Complete Program
The only bug in the this program is the mapping of the shortest path displaying an X every time you see a zero in the updated matrix.

Orginal matrix cannot contain many zeros or it will be mapped as x.

The alternative to fixing this one bug would be implementing a boolean and small number of that cell variable in the mapping while loop.

If number in (row + 1, col) is less than number (row, col +1) set boolean in nested for loop to true and set contents of (row + 1, col) to tiny. Reset boolean. (this requires extra if statements or just redoing the entire mapping loop)

Why a boolean? If the smaller number is 15 in (row + 1, col) and there is a 15 in (row, col + x) then you being doing another comparison.